==> 2.1-Recommend_User_Based_CF.ipynb <==
import numpy as np
import random
import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity
import pymongo
import json

#從xlsx匯入資料
df = pd.read_excel(r'./user_card_matrix_for_rec.xlsx',encoding='utf-16',index_col=0)

#mongodb匯入資料
client = pymongo.MongoClient(host='', port=27017)
#mongodb帳號密碼

client.admin.authenticate('', '')
db_rec = client.Recommend_card
coll_hold_card = db_rec.hold_card
user_card_matrix_for_rec=coll_hold_card.find()
df=pd.DataFrame(list(user_card_matrix_for_rec))
del df['_id']

# User_based_協同過濾推薦函數

def CF_Recommend_User_Based(my_card):
    #imput data type:my_card=['上海商銀-小小兵Bello回饋卡','中國信託-ANA聯名卡']
    hold_card_df=pd.DataFrame(columns=df.columns)
    hold_card_df.loc[0]=0
    #搜尋使用者持有卡於資料庫並設成1
    hold_card_df.loc[0,my_card]=1
    #計算cosine相似度
    user_sim_array=cosine_similarity(hold_card_df,df)
    sim_value_sort=pd.Series(user_sim_array[0]).sort_values(ascending = False)
    # df_m2m = pd.DataFrame(cosine_similarity(df))
    top3_sim_user={}
    
    
    for idx,v in sim_value_sort.iteritems():
        if len(top3_sim_user)<3:     #top3_sim_user
            top3_sim_user[v]=idx  
    weighted_card_rec=np.zeros(len(df.columns))
    #相似權重相加
    for i in top3_sim_user:
        weighted_card_rec+= df.loc[top3_sim_user[i]]*i  
    #未持有的卡推薦  
    recommend_card=[]
    weighted_card_rec_sort=weighted_card_rec.sort_values(ascending=False)
    for card,value in weighted_card_rec_sort.iteritems():
            if value >0:
                if hold_card_df[card][0] ==0 and len(recommend_card)<5:
                    recommend_card.append(card)
    if recommend_card:
        return recommend_card
    else:
        return '沒有好推薦的卡'

#測試推薦函數
CF_Recommend_User_Based(['國泰世華-KOKO COMBO icash聯名卡'])

# 以下函數整理及測試

my_card=['上海商銀-小小兵Bello回饋卡','中國信託-ANA聯名卡']
hold_card_df=pd.DataFrame(columns=df.columns)
hold_card_df.loc[0]=0
#搜尋使用者持有卡於資料庫並設成1
hold_card_df.loc[0,my_card]=1
user_sim_array=cosine_similarity(hold_card_df,df)
sim_value_sort=pd.Series(user_sim_array[0]).sort_values(ascending = False)

unique(1,1)


top3_sim_user={}
for idx,v in sim_value_sort.iteritems():
    if len(top3_sim_user)<3:
        top3_sim_user[v]=idx
        
    
    #top3_sim_user_idx.append(i.index)
print(top3_sim_user)
#print(top3_sim_user_idx)

weighted_card_rec=np.zeros(len(df.columns))
#相似權重相加
for i in top3_sim_user:
    weighted_card_rec+= df.loc[top3_sim_user[i]]*i
    
  


#未持有的卡推薦  
recommend_card=[]
weighted_card_rec_sort=weighted_card_rec.sort_values(ascending=False)
for card,value in weighted_card_rec_sort.iteritems():
        if value >0:
            if hold_card_df[card][0] ==0 and len(recommend_card)<5:
                recommend_card.append(card)
recommend_card

# 以下資料整理

def CF_Recommend_User_Based(my_card):
    #imput data type:my_card=['上海商銀-小小兵Bello回饋卡','中國信託-ANA聯名卡']
    hold_card_df=pd.DataFrame(columns=df.columns)
    hold_card_df.loc[0]=0
    #搜尋使用者持有卡於資料庫並設成1
    hold_card_df.loc[0,my_card]=1
    #計算cosine相似度
    user_sim_array=cosine_similarity(hold_card_df,df)
    # df_m2m = pd.DataFrame(cosine_similarity(df))
    top3_sim_user={}
    for idx,v in sim_value_sort.iteritems():
        if len(top3_sim_user)<3:     #top3_sim_user
            top3_sim_user[v]=idx  
    weighted_card_rec=np.zeros(len(df.columns))
    #相似權重相加
    for i in top3_sim_user:
        weighted_card_rec+= df.loc[top3_sim_user[i]]*i  
    #未持有的卡推薦  
    recommend_card=[]
    weighted_card_rec_sort=weighted_card_rec.sort_values(ascending=False)
    for card,value in weighted_card_rec_sort.iteritems():
            if value >0:
                if hold_card_df[card][0] ==0 and len(recommend_card)<5:
                    recommend_card.append(card)
    return(recommend_card)
    
    
    
    
    card_array = list(user_sim_array[0])
    #找出最相似的一筆，印出該筆資料中使用者未持有之卡片(值為0之卡片)
    # print(temp_c)
    max_sim_value=(max(card_array))
    max_sim_user=card_array.index(max_sim_value)
    # c_sort=sorted(temp_c,reverse=True)
    recommend_card=[]
    #找最相似人的持有卡
    for name_index, i in enumerate(df.loc[max_sim_user]):
        #如果最相似的人的卡在使用者未出現
        card=hold_card_df.columns[name_index]
        if i >0 and hold_card_df.loc[0,card] !=1:
            recommend_card.append(card)
    if recommend_card==[]:
        return '沒有好推薦的卡'
    else:
        return recommend_card

#測試推薦函數
CF_Recommend_User_Based(['台新銀行-@GoGo iCash御璽卡'])
#CF_Recommend(['花旗銀行-新世代花旗饗樂生活悠遊卡'])
#'台新銀行-@GoGo iCash御璽卡'
#'玉山銀行-U Bear信用卡'
#'永豐銀行-DAWHO現金回饋信用卡'
#中國信託-Yahoo聯名卡
#中國信託-MUJI無印良品聯名卡
#花旗銀行-新世代花旗饗樂生活悠遊卡

# 丟mongodb

#資料匯入mongoDB
#丟mongo
user_card_df=pd.read_excel(r'./user_card_matrix_for_rec.xlsx',encoding='utf-16',index_col=0)
user_card_df.head()

client = pymongo.MongoClient(host='123.241.175.34', port=27017)
client.admin.authenticate('root', '1qaz@WSX3edc')
db_rec = client.Recommend_card


coll_hold_card = db_rec.hold_card

hold_card= json.loads(user_card_df.T.to_json()).values()

coll_hold_card.insert_many(hold_card)
db_rec.list_collection_names()

#card180張資料匯入mongoDB
#丟mongo
user_card_df=pd.read_excel(r'./user_card_list.xlsx',encoding='utf-16')
user_card_df.head()

hold_card_list=json.loads(user_card_df.T.to_json()).values()

coll_hold_card_list = db_rec.hold_card_list

coll_hold_card_list.insert_many(hold_card_list)

db_rec.list_collection_names()

coll_hold_card_list = db_rec.hold_card_list
coll_hold_card_list_m=coll_hold_card_list.find()
df=pd.DataFrame(list(coll_hold_card_list_m))
df

# 以下皆資料整理

#df = df.drop(['Unnamed: 0'], axis=1)
df = df.drop(['總計'], axis=1)
#df = df.drop(['國泰世華-KOKO icash聯名卡.1'], axis=1)
#df = df.drop([2649], axis=0)
df=df.reset_index(drop=True)
df.fillna(0, inplace = True)
df.to_excel(r'./user_card_matrix_for_rec.xlsx',encoding='utf-16')

#df=df.drop(['Unnamed: 0'],axis=1)

df.columns

hold_card_df=pd.DataFrame(columns=df.columns)
hold_card_df.loc[0]=0
hold_card_df

CF_input_card(['上海商銀-小小兵Bello回饋卡'])

def CF_Recommend(my_card):
    #imput:my_card=['上海商銀-小小兵Bello回饋卡','中國信託-ANA聯名卡']
    a123=[]
    a=[]
    hold_card_df=pd.DataFrame(columns=df.columns)
    hold_card_df.loc[0]=0
    #搜尋使用者持有卡於資料庫並設成1
    hold_card_df.loc[0,my_card]=1
    #cosine
    c=cosine_similarity(hold_card_df,df)
    # df_m2m = pd.DataFrame(cosine_similarity(df))
    card_array = list(c[0])
    #找出最相似的一筆，印出該筆資料中使用者未持有之卡片(值為0之卡片)
    # print(temp_c)
    max_sim_value=(max(card_array))
    max_sim_user=card_array.index(max_sim_value)
    # c_sort=sorted(temp_c,reverse=True)
    recommend_card=[]
    #找最相似人的持有卡
    for name_index, i in enumerate(df.loc[max_sim_user]):
        #如果最相似的人的卡在使用者未出現
        card=hold_card_df.columns[name_index]
        if i >0 and hold_card_df.loc[0,card] !=1:
            recommend_card.append(card)
    if recommend_card==[]:
        return '沒有好推薦的卡'
    else:
        return recommend_card

def CF_Recommend(card_array):
    #找出最相似的一筆，印出該筆資料中使用者未持有之卡片(值為0之卡片)
    # print(temp_c)
    max_sim_value=(max(card_array))
    max_sim_user=card_array.index(max_sim_value)
    # c_sort=sorted(temp_c,reverse=True)
    recommend_card=[]
    #找最相似人的持有卡
    for name_index, i in enumerate(df.loc[max_sim_user]):
        #如果最相似的人的卡在使用者未出現
        card=hold_card_df.columns[name_index]
        if i >0 and hold_card_df.loc[0,card] !=1:
            recommend_card.append(card)
    if recommend_card==[]:
        return '沒有好推薦的卡'
    else:
        return recommend_card

card_array=CF_input_card([' 中國信託-MUJI無印良品聯名卡'])
CF_Recommend(card_array)

 for name_index, i in enumerate(df.loc[max_sim_user]):
        print(name_index,i)

CF_Recommend(card_array)

for n,i in enumerate(df.loc[max_sim_user]):
    print(i)

df.columns[0]

CF_Recommend(card_array)

card_array=CF_input_card(['上海商銀-小小兵Bello回饋卡'])

max_sim_value=(max(card_array))
max_sim_value

max_sim_user=card_array.index(max_sim_value)

df.loc[max_sim_user]

for n,i in enumerate(df.loc[max_sim_user]):
        if i >0 and hold_card_df[n][0]!=1:
            print(df.columns[n])

for n,i in enumerate(df.loc[max_sim_user]):
    print(n,i)
