==> 2.2-Recommend_ITEM_Based_CF.ipynb <==
import numpy as np
import random
import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity
import pymongo

#從xlsx匯入資料
df = pd.read_excel(r'./user_card_matrix_for_rec.xlsx',encoding='utf-16',index_col=0)

#從mongo匯入資料
client = pymongo.MongoClient(host='', port=)
#mongodb帳號密碼

client.admin.authenticate('', '')
db_rec = client.Recommend_card
coll_hold_card = db_rec.hold_card
user_card_matrix_for_rec=coll_hold_card.find()
df=pd.DataFrame(list(user_card_matrix_for_rec))
del df['_id']

# Item_Based_協同過濾推薦函數

def CF_Recommend_Item_Based(my_card):
    #imput data type:my_card=['上海商銀-小小兵Bello回饋卡','中國信託-ANA聯名卡']
    #轉置df
    df_T=df.T
    #計算各卡間的相似度
    card_sim=cosine_similarity(df_T,df_T)
    #找出已持有卡index
    indices = pd.Series(df_T.index)   #所有卡series
    card_index_list=[indices[indices == name].index[0] for name in my_card]  #找出各卡index
    weighted_card_rec=np.zeros(len(df_T.index))   #創全0且為卡片數量長度的array
    for  i in card_index_list:
        weighted_card_rec +=card_sim[i]           #將持有卡於其他卡的相似度相加成推薦權重
    #排列相似度最高前5名卡片
    weighted_score=pd.Series(weighted_card_rec).sort_values(ascending = False)
    top_5_indexes=weighted_score.iloc[len(my_card):len(my_card)+5].index
    rec_card=[]
    for i in top_5_indexes:
        rec_card.append(indices[i])
    return rec_card

#測試函數
#測試推薦函數
CF_Recommend_Item_Based(['國泰世華-KOKO COMBO icash聯名卡'])
#CF_Recommend(['花旗銀行-新世代花旗饗樂生活悠遊卡'])
#'台新銀行-@GoGo iCash御璽卡'
#'玉山銀行-U Bear信用卡'
#'永豐銀行-DAWHO現金回饋信用卡'
#中國信託-Yahoo聯名卡
#中國信託-MUJI無印良品聯名卡
#花旗銀行-新世代花旗饗樂生活悠遊卡

# 以下皆測試

df=df.T

df.head()

hold_card_df=pd.DataFrame(columns=df.columns)
hold_card_df.loc[0]=0
my_card=['上海商銀-小小兵Bello回饋卡','中國信託-ANA聯名卡']
hold_card_df.loc[0,my_card]=1

hold_card_df=hold_card_df.T

df_combine=pd.concat([df,hold_card_df]).reset_index(drop=True)

df_combine

df_combine_T=df_combine.T

df_combine_T.head()

cosine_similarity(df_combine_T,df_combine_T).shape

indices = pd.Series(df_T.index)
indices[indices == '三信銀行-VISA白金卡'].index[0]

indices[1]

my_card=['上海商銀-小小兵Bello回饋卡','中國信託-ANA聯名卡']

idx=[indices[indices == name].index[0] for name in my_card]
idx

df_T=df.T
card_sim=cosine_similarity(df_T,df_T)
    #找出已持有卡
indices = pd.Series(df_T.index)
my_card=['上海商銀-小小兵Bello回饋卡','中國信託-ANA聯名卡']

card_index_list=[indices[indices == name].index[0] for name in my_card]

weighted_card_rec=np.zeros(len(df_T.index))

for  i in card_index_list:
    weighted_card_rec +=card_sim[i]

weighted_score=pd.Series(weighted_card_rec).sort_values(ascending = False)
top_5_indexes=weighted_score.iloc[len(my_card)-1:len(my_card)+4].index
rec_card=[]
for i in top_5_indexes:
    rec_card.append(indices[i])

CF_Recommend_item_based(['上海商銀-小小兵Bello回饋卡','中國信託-ANA聯名卡'])
