==> 3.1-GPS_data_ETL_Raw2df.ipynb <==
import os
import gpxpy
import re
import glob
import pandas as pd
import pytz
import time
import datetime

dir_path=r'E:/專題/credit_card/DB104_hau/GPS/data'

# gpx_to_dict

dir_path='E:/專題/credit_card/DB104_hau/GPS/data'
results = [os.path.basename(f) for f in glob.glob(os.path.join(dir_path, '*.gpx'))]
all_gps_record=[]
localFormat = "%Y-%m-%d %H:%M:%S"
for name in results:
    with open(r'./data/{}'.format(name),'r',encoding='utf-8') as gdata:
        gpx = gpxpy.parse(gdata)
    for track in gpx.tracks:
                for segment in track.segments:
                    for point in segment.points:
                        time = ""
                        if point.time:
                            time = point.time.astimezone(pytz.timezone('Asia/Taipei')).strftime(localFormat)
                        gps_record={'user':name[:5],'time':time,'latitude':point.latitude,'longitude':point.longitude,'elevation':point.elevation,\
                                    'speed':point.speed,'source':point.source,'horizontal_dilution':point.horizontal_dilution,\
                                    'vertical_dilution':point.vertical_dilution,'course':point.course,'satellites':point.satellites}
                        all_gps_record.append(gps_record)

col_name=['user','time','latitude','longitude','elevation','speed','source','horizontal_dilution','vertical_dilution','course','satellites']
user_track_df=pd.DataFrame(all_gps_record,columns=col_name)
user_track_df.head()

user_track_df['speed']=user_track_df['speed']*3.6  #速度單位轉換m/s-->km/hr

user_track_df.head()

user_track_df.to_excel(r'user_gps_data.xlsx',encoding='utf-16')

# 匯入DF

user_gps_df=pd.read_excel(r'./user_gps_data.xlsx',encoding='utf-16')

# 時區調整

point.time

point.time.astimezone(pytz.timezone('Asia/Taipei'))

# 找出所有檔名

#[os.path.basename(f) for f in glob.glob(os.path.join(dir_path, '*.gpx'))]
#glob.glob(dir_path)
#glob.glob(os.path.join(dir_path, '*.gpx'))

localFormat = "%Y-%m-%d %H:%M:%S"
localDatetime = point.time.astimezone(pytz.timezone('Asia/Taipei')).strftime(localFormat)
print(localDatetime.strftime(localFormat))


with open(r'./data/user7_20191129.gpx','r',encoding='utf-8') as gdata:
    gpx = gpxpy.parse(gdata)
    points2 = list()
    for track in gpx.tracks:
        for segment in track.segments:
            for point in segment.points:
                print(point)
#                 dict = {'Timestamp': point.time,
#                         'Latitude': point.latitude,
#                         'Longitude': point.longitude,
#                         'Elevation': point.elevation,
#                         'Accuracy':point.accuracy
#                         }
#                 points2.append(dict)
# print(points2)

with open(r'./data/user1_20191126.gpx','r',encoding='utf-8') as gdata:
    gpx = gpxpy.parse(gdata)
print(gpx.tracks)

all_gps_record[2000:2100]

dir_path='E:/專題/credit_card/DB104_hau/GPS/data'
results = [os.path.basename(f) for f in glob.glob(os.path.join(dir_path, '*.gpx'))]
for i in results:
    print(i[:5])
