==> 3.2-User_GPS_to_Stay_Point.ipynb <==
import pandas as pd
from geopy.distance import geodesic
from datetime import datetime
import gmplot
import folium
import numpy as np
#import time

# 兩點距離求速度_1

gps_data_df=pd.read_excel(r'./user_gps_data1.xlsx',encoding='utf-16',index_col=0)

localFormat = "%Y-%m-%d %H:%M:%S"
gps_data_df['point_speed']=0
gps_data_df['point_distance']=0
for user in list(set(gps_data_df['user'])):
    for i in gps_data_df[gps_data_df['user']==user].index:
        if i < gps_data_df[gps_data_df['user']==user].index[-1]:
            deltatime=datetime.strptime(gps_data_df.loc[i+1,'time'],localFormat)-datetime.strptime(gps_data_df.loc[i,'time'],localFormat)
            minutesdelta=(deltatime.seconds)/60
            if minutesdelta<10:
                prev=gps_data_df.loc[i,['latitude','longitude']]
                post=gps_data_df.loc[i+1,['latitude','longitude']]
                if minutesdelta>0:
                    gps_data_df.loc[i+1,'point_speed']=((geodesic(prev,post).kilometers)/minutesdelta)*60
                gps_data_df.loc[i+1,'point_distance']=(geodesic(prev,post).kilometers)
    #print(prev,post)
gps_data_df.head()

gps_data_df[['user','time','latitude','longitude','point_speed']].head()

# 求staypoint_2

#求staypoint
gps_data_df=pd.read_excel(r'./gps_data_pointspeed2.xlsx',encoding='utf-16',index_col=0)
#門檻值
#3km/hr 15mins or #1.5km/hr 15mins
speed_threshold=3
stay_point_user_list={}
gps_data_df['stay_point_lat']=0
gps_data_df['stay_point_long']=0
localFormat = "%Y-%m-%d %H:%M:%S"
for user in list(set(gps_data_df['user'])):    
    stay_point_coor_lat=[]
    stay_point_coor_lon=[]
    stay_time=0
    stay_point_list=[]
    stay_point_index=[]
    for i in gps_data_df[gps_data_df['user']==user].index[1:]:
                #if speed <3
                if gps_data_df.loc[i,'point_speed']<speed_threshold:
                    #設定每個點時間，加總每個點時間
                    #total_time+=time_delta
                    deltatime=datetime.strptime(gps_data_df.loc[i,'time'],localFormat)-datetime.strptime(gps_data_df.loc[i-1,'time'],localFormat)
                    minutesdelta=(deltatime.seconds)/60
                    if minutesdelta<15:
                        stay_time +=minutesdelta
                        #stay point.append(coord)
                        stay_point_coor_lat.append(gps_data_df.loc[i,'latitude'])
                        stay_point_coor_lon.append(gps_data_df.loc[i,'longitude'])
                        stay_point_index.append(i)
                    else:
                        if stay_point_coor_lat:
                            stay_point=(sum(stay_point_coor_lat)/len(stay_point_coor_lat),
                                        sum(stay_point_coor_lon)/len(stay_point_coor_lon))
                            gps_data_df.loc[stay_point_index,'stay_point_lat']=stay_point[0]
                            gps_data_df.loc[stay_point_index,'stay_point_long']=stay_point[1]
                        stay_point_list.append(stay_point)
                        stay_point_coor_lat=[]
                        stay_point_coor_lon=[]
                        stay_point_index=[]
                        stay_time=0
                        #stay point.append(coord)


                    #if total_time>10 minutes

                elif gps_data_df.loc[i,'point_speed'] > speed_threshold and stay_time>15:
                        stay_point=(sum(stay_point_coor_lat)/len(stay_point_coor_lat),
                                    sum(stay_point_coor_lon)/len(stay_point_coor_lon))
                        gps_data_df.loc[stay_point_index,'stay_point_lat']=stay_point[0]
                        gps_data_df.loc[stay_point_index,'stay_point_long']=stay_point[1]
                        stay_point_list.append(stay_point)
                        stay_point_coor_lat=[]
                        stay_point_coor_lon=[]
                        stay_point_index=[]
                        stay_time=0
                else:
                    stay_point_coor_lat=[]
                    stay_point_coor_lon=[]
                    stay_point_index=[]
                    stay_time=0
    stay_point_user_list[user]=stay_point_list
                
#print(stay_point_user_list)
                

gps_data_df.to_excel(r'./gps_data_staypoint.xlsx',encoding='utf-16')

# 每個位置合併成一個_3

#座標合併成1格 & 只取日期
stay_point_df=pd.read_excel(r'./gps_data_staypoint3.xlsx',encoding='utf-16',index_col=0)
subset = stay_point_df[['stay_point_lat','stay_point_long']]
subset
stay_point_df['date']=[x[:10] for x in stay_point_df['time'].values]
stay_point_df['stay_point_coord']=[tuple(x) for x in subset.to_numpy()]
#del stay_point_df['coord']

#求每點間隔時間
stay_point_df_d=stay_point_df

stay_point_df_d['delta_time']=0

stay_point_df_d.head()

localFormat = "%Y-%m-%d %H:%M:%S"
for user in list(set(stay_point_df_d['user'])):    
    for i in stay_point_df_d[stay_point_df_d['user']==user].index[1:]:
                    #設定每個點時間，加總每個點時間
                    #total_time+=time_delta
                    deltatime=datetime.strptime(stay_point_df_d.loc[i,'time'],localFormat)-datetime.strptime(stay_point_df_d.loc[i-1,'time'],localFormat)
                    minutesdelta=(deltatime.seconds)/60
                    if minutesdelta<15:
                        stay_point_df_d.loc[i,'delta_time']=minutesdelta
                        #print(stay_point_df_d.loc[i,'delta_time'])


stay_point_df_d.to_excel(r'./stay_point_df_complete.xlsx',encoding='utf=16')

## 求stay_point 開始及結束時間，group成1組_4

#求stay_point 開始及結束時間


stay_point_df_complete=pd.read_excel(r'./stay_point_df_complete4.xlsx',encoding='utf-16',index_col=0)
stay_point_group_df_time=stay_point_df_complete[['user','date','time','stay_point_coord']]
group_df=stay_point_group_df_time.groupby(['user','date','stay_point_coord'])
list(group_df)
group_df.agg([np.min,np.max])
#
localFormat = "%Y-%m-%d %H:%M:%S"
groupby_time=group_df.agg([np.min,np.max])
groupby_time.columns

tmp=[]
for x,n in zip(groupby_time['time']['amax'].values,groupby_time['time']['amin'].values):
    deltatime=(datetime.strptime(x,localFormat)-datetime.strptime(n,localFormat)).seconds/60
    tmp.append(deltatime)

groupby_time['duration']=pd.Series(tmp).values


groupby_time.to_excel(r'./group_time.xlsx',encoding='utf-16')

#求兩stay_point中間 最大速度及開始結束時間

#求起訖時間
stay_point_group_df_time=stay_point_df_complete[['user','date','time','stay_point_coord']]
group_df=stay_point_group_df_time.groupby(['user','date','stay_point_coord'])
list(group_df)
group_df.agg([np.min,np.max])

groupby_time=pd.read_excel(r'./group_time.xlsx',encoding='utf-16')

#
localFormat = "%Y-%m-%d %H:%M:%S"
groupby_time=group_df.agg([np.min,np.max])
groupby_time.columns
#deltatime=datetime.strptime(groupby_time['time']['amax'].values,localFormat)-datetime.strptime(groupby_time['time']['amin'].values,localFormat)
#print(deltatime)
#minutesdelta=(deltatime.seconds)/60
tmp=[]
for x,n in zip(groupby_time['time']['amax'].values,groupby_time['time']['amin'].values):
    deltatime=(datetime.strptime(x,localFormat)-datetime.strptime(n,localFormat)).seconds/60
    tmp.append(deltatime)

groupby_time['duration']=pd.Series(tmp).values
    #groupby_time.loc[i,'duration']=
#-groupby_time['time']['amin']
#groupby_time['duration']=groupby_time['duration']


#.append(pd.Series(tmp),ignore_index=True)
#pd.Series(tmp)
#groupby_time.sort_values(by=['amin'])

stay_point_df_d=stay_point_df

# 求每點時間

groupby_time.to_excel(r'./group_time.xlsx',encoding='utf-16')

groupby_time

# 相鄰stay_point距離小於50m合併_5

user_stay_point_daily=pd.read_excel(r'./user_stay_point_daily5.xlsx',encoding='utf-16')
#調整stay_point位置
#若下個staypoint距離前個<30m  視同同一個staypoint
for u in set(user_stay_point_daily['user']):
    date_set=set(user_stay_point_daily[user_stay_point_daily['user']==u]['date'])
    user_index=user_stay_point_daily[user_stay_point_daily['user']==u].index
    user_df=user_stay_point_daily.loc[user_index]
    for d in date_set:
        user_date_index=user_df[user_df['date']==d].index
        for i in user_date_index:
            if i+1 in user_date_index:
                now=eval(user_stay_point_daily.loc[i,'stay_point_coord'])
                nextp=eval(user_stay_point_daily.loc[i+1,'stay_point_coord'])
                dist=geodesic(now,nextp).kilometers
                if dist<0.05:
                    user_stay_point_daily.loc[i+1,'stay_point_coord']=user_stay_point_daily.loc[i,'stay_point_coord']
    #for d in date_set:
       
       #user_stay_point_daily[user_stay_point_daily['user']==u and user_stay_point_daily['date']==d]
user_stay_point_daily    

#groupby 開始時間及duration
user_stay_point_adjust=user_stay_point_daily
adjust_group=user_stay_point_adjust.groupby(['user','date','stay_point_coord']).agg({'start':'min',
                                                                       'end':'max',
                                                                       'duration':'sum'})

#break multiindex###########
adjust_group.reset_index(inplace=True)
#組內依starttime排序
adjust_group_2=adjust_group.groupby(['user','date']).apply(lambda x:x.sort_values('start'))

adjust_group_2.to_excel(r'./stay_point_nearbygroup.xlsx',encoding='utf-16')

adjust_group=pd.read_excel(r'./stay_point_nearbygroup.xlsx',encoding='utf-16')

#調整duration時間
localFormat = "%Y-%m-%d %H:%M:%S"
for i in adjust_group.index:
    start=adjust_group.loc[i,'start']
    end=adjust_group.loc[i,'end']
    deltatime=(datetime.strptime(str(end),localFormat)-datetime.strptime(str(start),localFormat)).seconds/60
    adjust_group.loc[i,'duration']=deltatime
adjust_group

adjust_group.to_excel(r'./user_stay_point_final.xlsx',encoding='utf-16')

# #求速度

# for user 
# for i in index
# speed_list=[]
# if stay_point ==(0,0) and point_speed>0:
#     #df.loc[i,'end_time']=df.loc[i,'time']
#     speed_list.append(df.loc[i,'point_speed'])    
# else:
#     if speed_list:
#         df.loc[i]['move_max_speed']=max(speed_list)
#         df.loc[i,'start_time']=df.loc[i,'time']
#         speed_list=[]
