==> 3.3-Calculate_User_Behavior.ipynb <==
import googlemaps
import pandas as pd
import time
from geopy.distance import geodesic
import numpy as np

# 計算使用者去電影院次數

cinema_list=pd.read_excel(r'./cinema_list.xlsx',encoding='utf-16',index_col=0)

cinema_list.head()

user_stay_point=pd.read_excel(r'./user_stay_point_final.xlsx',encoding='utf-16',index_col=0)
user_stay_point.head()

#電影院座標轉換成1格
coord_list=[]
#cinema_list['coord']=(0,0)
for i in range(len(cinema_list)):
    coord=(cinema_list.loc[i,'latitude'],cinema_list.loc[i,'longtiude'])
    coord_list.append(coord)
pds=pd.Series(coord_list)
cinema_list['coord']=pds

user_movie={}
for i in user_stay_point.index:
    if user_stay_point.loc[i,'user'] not in user_movie:
                  user_movie[user_stay_point.loc[i,'user']]=0
    stay_point=eval(user_stay_point.loc[i,'stay_point_coord'])
    for c in cinema_list['coord'].index:
        dist=geodesic(stay_point,cinema_list['coord'][c]).kilometers
        if dist <0.1 and user_stay_point.loc[i,'duration']>60 :
            print(cinema_list.loc[c])
            print(user_stay_point.loc[i])
            user_movie[user_stay_point.loc[i,'user']]+=1


user_movie

# 計算使用者平均staypoint個數及停留時間

user_stay_point.groupby(['user']).agg({'stay_point_coord':'count',
                                      'duration':'sum'})

user_stay_point
