==> 4-Nearby_Card_discount_store.ipynb <==
from geopy.distance import geodesic
from geopy.geocoders import Nominatim
import googlemaps
import requests
import pandas as pd
from bs4 import BeautifulSoup
import json
import time

df_store_list=pd.read_excel(r'./store_location.xlsx',encoding='utf-16',index_col=0)

# df_store_list=df_store_list.reset_index()
# df_store_list=df_store_list.drop(['index'],axis=1)

df_store_list.head()

#設計回傳附近店家優惠函數
def near_by_info(lat,lng):
    #import googlemaps
#     輸入API憑證，上限40000次
#     gmap=googlemaps.Client(key='')
#     #轉換地址為經緯度
#     geocode=gmap.geocode(now_place)
#     #若無法轉換成功
#     if geocode==[]:
#         print('偵測不到位置')

    
#     else:
    #code=geocode[0]['geometry']['location']
    #now_code=(code['lat'],code['lng'])
    now_code=(lat,lng)
    store=[]
    addr=[]
    info=[]
    distance=[]

        #找尋資料庫df_store_list中商家位置，計算距離
    for i in range(len(df_store_list)):
        address=(df_store_list['lat'][i],df_store_list['lng'][i])
        #計算當下位置與商家位置距離
        dist=geodesic(address,(lat,lng)).kilometers
        #若小於5公里
        if dist<5:
            store.append(df_store_list['store'][i])
            if df_store_list['store'][i]=='屈臣氏':
                info.append('刷LINEPay卡5%回饋')
            elif df_store_list['store'][i]=='美廉社':
                info.append('刷LINEPay卡2%回饋')
            addr.append(df_store_list['address'][i])
            distance.append(str(round(dist,2))+'km')
    neardf=pd.DataFrame({'店家':store,'優惠內容':info,'地址':addr,'距離':distance},columns=['店家','優惠內容','地址','距離'])

    if len(neardf)==0:
        print('附近沒有優惠店家')
    else:
        return neardf.sort_values(by=['距離'])[:5]


near_by_info(24.959190, 121.211006)
